+++ include/lib/elf/elf_helpers.h
@@ -9,18 +9,28 @@
 #include <string.h>
 #include <types.h>
 
-#include <elf.h>
+#include <lib/elf/elf.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif// __cplusplus
 
-static inline u8 fw_elf_get_class(u32 elf_fw_addr) {
+/**
+ * @brief Read the ELF class from a firmware image.
+ * @param[in] elf_fw_addr Address of the ELF image header.
+ * @return ELFCLASS32 or ELFCLASS64.
+ */
+static inline u8 fw_elf_get_class(uintptr_t elf_fw_addr) {
 	struct elf32_hdr *ehdr = (struct elf32_hdr *) elf_fw_addr;
 
 	return ehdr->e_ident[EI_CLASS];
 }
 
+/**
+ * @brief Initialize the identification bytes in an ELF header.
+ * @param[out] hdr ELF header to initialize.
+ * @param[in] class ELFCLASS32 or ELFCLASS64.
+ */
 static inline void elf_hdr_init_ident(struct elf32_hdr *hdr, u8 class) {
 	memcpy(hdr->e_ident, ELFMAG, SELFMAG);
 	hdr->e_ident[EI_CLASS] = class;
